+++ mmaction/datasets/__init__.py
@@ -4,6 +4,7 @@ from .audio_feature_dataset import AudioFeatureDataset
 from .base import BaseDataset
 from .builder import build_dataloader, build_dataset
 from .dataset_wrappers import RepeatDataset
+from .hvu_dataset import HVUDataset
 from .rawframe_dataset import RawframeDataset
 from .ssn_dataset import SSNDataset
 from .video_dataset import VideoDataset
@@ -11,5 +12,5 @@ from .video_dataset import VideoDataset
 __all__ = [
     'VideoDataset', 'build_dataloader', 'build_dataset', 'RepeatDataset',
     'RawframeDataset', 'BaseDataset', 'ActivityNetDataset', 'SSNDataset',
-    'AudioDataset', 'AudioFeatureDataset'
+    'HVUDataset', 'AudioDataset', 'AudioFeatureDataset'
 ]