+++ mmaction/models/recognizers/base.py
@@ -40,6 +40,14 @@ class BaseRecognizer(nn.Module, metaclass=ABCMeta):
 
         self.train_cfg = train_cfg
         self.test_cfg = test_cfg
+
+        # aux_info is the list of tensor names beyond 'imgs' and 'label' which
+        # will be used in train_step and val_step, data_batch should contain
+        # these tensors
+        self.aux_info = []
+        if train_cfg is not None and 'aux_info' in train_cfg:
+            self.aux_info = train_cfg['aux_info']
+
         self.init_weights()
 
         self.fp16_enabled = False
@@ -100,7 +108,7 @@ class BaseRecognizer(nn.Module, metaclass=ABCMeta):
         return cls_score
 
     @abstractmethod
-    def forward_train(self, imgs, labels):
+    def forward_train(self, imgs, labels, **kwargs):
         """Defines the computation performed at every call when training."""
         pass
 
@@ -146,14 +154,14 @@ class BaseRecognizer(nn.Module, metaclass=ABCMeta):
 
         return loss, log_vars
 
-    def forward(self, imgs, label=None, return_loss=True):
+    def forward(self, imgs, label=None, return_loss=True, **kwargs):
         """Define the computation performed at every call."""
         if return_loss:
             if label is None:
                 raise ValueError('Label should not be None.')
-            return self.forward_train(imgs, label)
+            return self.forward_train(imgs, label, **kwargs)
         else:
-            return self.forward_test(imgs)
+            return self.forward_test(imgs, **kwargs)
 
     def train_step(self, data_batch, optimizer, **kwargs):
         """The iteration step during training.
@@ -184,7 +192,12 @@ class BaseRecognizer(nn.Module, metaclass=ABCMeta):
         imgs = data_batch['imgs']
         label = data_batch['label']
 
-        losses = self(imgs, label)
+        aux_info = {}
+        for item in self.aux_info:
+            assert item in data_batch
+            aux_info[item] = data_batch[item]
+
+        losses = self(imgs, label, return_loss=True, **aux_info)
 
         loss, log_vars = self._parse_losses(losses)
 
@@ -205,7 +218,11 @@ class BaseRecognizer(nn.Module, metaclass=ABCMeta):
         imgs = data_batch['imgs']
         label = data_batch['label']
 
-        losses = self(imgs, label)
+        aux_info = {}
+        for item in self.aux_info:
+            aux_info[item] = data_batch[item]
+
+        losses = self(imgs, label, return_loss=True, **aux_info)
 
         loss, log_vars = self._parse_losses(losses)